PIM-4980: move flat normalizers in Flat directory
