+++ src/Pim/Bundle/LocalizationBundle/Resources/config/normalizers.yml
@@ -1,30 +1,31 @@
 parameters:
-    pim_localization.normalizer.metric.class:        Pim\Component\Localization\Normalizer\MetricNormalizer
-    pim_localization.normalizer.price.class:         Pim\Component\Localization\Normalizer\PriceNormalizer
-    pim_localization.normalizer.product_value.class: Pim\Component\Localization\Normalizer\ProductValueNormalizer
+    pim_localization.normalizer.flat.metric.class:        Pim\Component\Localization\Normalizer\Flat\MetricNormalizer
+    pim_localization.normalizer.flat.price.class:         Pim\Component\Localization\Normalizer\Flat\PriceNormalizer
+    pim_localization.normalizer.flat.product_value.class: Pim\Component\Localization\Normalizer\Flat\ProductValueNormalizer
 
 services:
-    pim_localization.normalizer.metric:
+    # flat normalizers
+    pim_localization.normalizer.flat.metric:
         public: false
-        class: %pim_localization.normalizer.metric.class%
+        class: %pim_localization.normalizer.flat.metric.class%
         arguments:
             - '@pim_serializer.normalizer.flat.metric'
             - '@pim_localization.localizer.metric'
         tags:
             - { name: pim_serializer.normalizer, priority: 95 }
 
-    pim_localization.normalizer.price:
+    pim_localization.normalizer.flat.price:
         public: false
-        class: %pim_localization.normalizer.price.class%
+        class: %pim_localization.normalizer.flat.price.class%
         arguments:
             - '@pim_serializer.normalizer.flat.price'
             - '@pim_localization.localizer.price'
         tags:
             - { name: pim_serializer.normalizer, priority: 95 }
 
-    pim_localization.normalizer.product_value:
+    pim_localization.normalizer.flat.product_value:
         public: false
-        class: %pim_localization.normalizer.product_value.class%
+        class: %pim_localization.normalizer.flat.product_value.class%
         arguments:
             - '@pim_serializer.normalizer.flat.product_value'
             - '@pim_localization.localizer.registry'
